Refactor camera helper
